pyproject.toml
[project]
name = "wechat-article-collect"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "langgraph",
    "langchain",
    "langchain-openai",
    "openpyxl",
    "pandas",
    "beautifulsoup4",
    "python-dateutil",
    "python-dotenv"
]
